@@ -41,19 +41,19 @@
     "d3": "3.5.10",
     "dotenv": "1.2.0",
     "fixed-data-table": "0.6.0",
-    "font-awesome": "4.4.0",
+    "font-awesome": "~4.4.0",
     "history": "~1.13.1",
     "lodash": "3.10.1",
     "moment": "2.10.6",
     "numeral": "1.5.3",
     "nvd3": "1.8.1",
-    "react": "0.14.0",
+    "react": "0.14.3",
     "react-dimensions": "1.0.1",
     "react-dom": "0.14.3",
     "react-nvd3": "0.3.0",
     "react-popout": "0.4.0",
     "react-router": "~1.0.2",
-    "react-sparklines": "1.4.2",
+    "react-sparklines": "~1.4.2",
     "rx": "4.0.7",
     "yargs": "3.18.0"
   },
@@ -67,9 +67,9 @@
     "chalk": "1.1.1",
     "copy-webpack-plugin": "0.3.3",
     "css-loader": "0.23.0",
-    "eslint": "1.0.0",
+    "eslint": "1.10.3",
     "eslint-config-airbnb": "2.1.1",
-    "eslint-plugin-react": "3.3.1",
+    "eslint-plugin-react": "3.12.0",
     "extract-text-webpack-plugin": "0.9.0",
     "file-loader": "0.8.4",
     "html-webpack-plugin": "1.6.1",
